Bring back tsGenerateFunctionFactoryCodeForRealmJSAndWebV0:
I wanted to implement this in js-and-web-runtime-and-rollup-plugins but the problem is
that for this function to work properly it needs things from the typescript package.

It makes more sense to implement it here since this package is meant to abstract
all of typescript away.

--

The aliases are not really needed either since it makes only sense to refer to
AnioJsDependencies from within either #~src or #~auto/.